+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Add support for PHP v7.1 (temporary measure to move away from PHP v5).
 - Type declarations have been added to all method parameters and return types
   where possible.
+- Constant `Consistent::HASH_AVAILABLE` to make it easier to extend the class to
+  add alternate hash methods.
 ### Removed
 - **BC break**: Removed support for PHP versions < v7.1 as they are no longer
   [actively supported](https://php.net/supported-versions.php) by the PHP project.